Changed seqTotals and added renameBathFastq
Changed lin 29 of seqTotals.py, removing the "F" on the regex expression to create the sampleName. This way, the function can search for any string and replace it, instead of simply looking for a string starting with "F". Hopefully this makes it easier to apply to .html files with different names.
Added the renameBatchFastq.py script. This script takes gziped fastq (.fq.gz) files from all present subdirectories and renames them with subdirectory name + sequence identifier (_1 for FW, _2 for RV).